upload

+++ app.py
@@ -143,11 +143,15 @@ class ConnectToMySQL():
 			if self.con:
 				self.con.close()
 
-	def UploadData(self, lineEdit_id, lineEdit_login, lineEdit_password, lineEdit_fio, lineEdit_pol, lineEdit_data, lineEdit_addres, label_status):
+	def UploadData(self, lineEdit_id, lineEdit_login, lineEdit_password, lineEdit_fio, lineEdit_pol, lineEdit_data, lineEdit_addres, label_status, is_admin, is_programm_user):
 		try:
 			self.connect()
 			cursor = self.con.cursor()
 			sql = f"INSERT INTO user (`id`, `login`, `password`, `fio`, `pol`, `birthday`, `address`) VALUES ('{lineEdit_id}', '{lineEdit_login}', '{lineEdit_password}', '{lineEdit_fio}', '{lineEdit_pol}', '{lineEdit_data}', '{lineEdit_addres}');"
+			cursor.execute(sql)
+			self.con.commit()
+
+			sql = f"INSERT INTO prava (`User_id`, `is_admin`, `is_programm_user`) VALUES ('{lineEdit_id}', '{is_admin}', '{is_programm_user}');"
 			cursor.execute(sql)
 			self.con.commit()
 			
@@ -203,30 +207,40 @@ class Authorization(QtWidgets.QWidget):
 		rules = self.connect.rules()
 		
 		for item in result:
-			if (not self.user_login) or (not self.user_password):
-				self.status.setText('Вы не ввели логин или пароль')
-			else:
-				for i in rules:
-					if (str(self.user_login) == str(item['login']) and str(self.user_password) == str(item['password']) and (item['id'] == i['User_id'] and i['is_programm_user'] == 1)):
-						self.status.clear()
-						
-						self.username = item['login']
-						
-						self.window = MainWindow()
-						window.label_user.setText(self.username)
-						
-						
-						if (item['id'] == i['User_id'] and i['is_admin'] == 1):
-							window.btn_add_1.show()
-							window.btn_add_2.show()
-						else:
+			for i in rules:
+
+				self.username = item['login']
+				self.window = MainWindow()
+
+				if (not self.user_login) or (not self.user_password):
+					self.status.setText('Вы не ввели логин или пароль')
+				elif (str(self.user_login) != str(item['login']) or str(self.user_password) != str(item['password']) ):
+					self.status.setText('Введен неправильно логин или пароль')
+				else:
+					if (str(self.user_login) == str(item['login']) and str(self.user_password) == str(item['password']) and (item['id'] == i['User_id'])):
+						if (item['id'] == i['User_id'] and i['is_programm_user'] == 1):
+							self.status.clear()
+
 							window.btn_add_1.hide()
 							window.btn_add_2.hide()
-					
-						self.close()
-					elif (str(self.user_login) != str(item['login']) and str(self.user_password) == str(item['password']) or str(self.user_login) == str(item['login']) and str(self.user_password) != str(item['password'])):
-						self.status.setText('Вы ввели не верный логин или пароль')
-						window.label_user.clear()
+
+							window.label_user.setText(self.username)
+
+							self.close()
+						
+						if (item['id'] == i['User_id'] and i['is_admin'] == 1):
+							self.status.clear()
+
+							window.btn_add_1.show()
+							window.btn_add_2.show()
+
+							window.label_user.setText(self.username)
+
+							self.close()
+
+						break
+
+
 
 class MainWindow(QtWidgets.QMainWindow):
 	def __init__(self):
@@ -260,6 +274,8 @@ class MainWindow(QtWidgets.QMainWindow):
 		self.btn_add_2 = self.appUI.btn_add_2
 		self.btn_add_1.hide()
 		self.btn_add_2.hide()
+		self.btn_logout_1 = self.appUI.pushButton_4
+		self.btn_logout_2 = self.appUI.pushButton_8
 
 		self.label_user = self.appUI.label_user
 		
@@ -268,8 +284,8 @@ class MainWindow(QtWidgets.QMainWindow):
 
 		self.appUI.status_label_2.hide()
 
-		self.result_table_2.hide()
-		self.appUI.widget_10.hide()
+		# self.result_table_2.hide()
+		# self.appUI.widget_10.hide()
 
 		self.result_table_1.horizontalHeader().setSectionResizeMode(QHeaderView.ResizeToContents)
 		self.result_table_1.setEditTriggers(QtWidgets.QAbstractItemView.NoEditTriggers)
@@ -294,6 +310,8 @@ class MainWindow(QtWidgets.QMainWindow):
 		self.btn_add_photo.clicked.connect(self.on_btn_add_photo)
 		self.bnt_start_process.clicked.connect(self.on_bnt_start_process)	
 		self.btn_user.clicked.connect(self.on_btn_user)
+		self.btn_logout_1.clicked.connect(self.btn_logout)
+		self.btn_logout_2.clicked.connect(self.btn_logout)
 
 	def getsamerowcell(self, columnname, table):
 
@@ -349,6 +367,21 @@ class MainWindow(QtWidgets.QMainWindow):
 
 		fname = QFileDialog.getOpenFileName(self, 'Open file', "")
 		return fname[0]
+
+	@pyqtSlot(bool)
+	def btn_logout(self):
+		self.label_user.clear()
+		self.btn_add_1.hide()
+		self.btn_add_2.hide()
+		self.on_btn_clear_table_1()
+		self.on_btn_clear_table_2()
+		self.appUI.label_13.clear()
+		self.appUI.label_15.clear()
+		self.appUI.label_17.clear()
+		self.appUI.label_19.clear()
+		self.appUI.label_21.clear()
+		if (self.appUI.stackedWidget.currentIndex() == 2):
+			self.appUI.stackedWidget.setCurrentIndex(0)
 	
 	@pyqtSlot(bool)
 	def on_btn_user(self):
@@ -357,82 +390,90 @@ class MainWindow(QtWidgets.QMainWindow):
 		
 	@pyqtSlot(bool)
 	def on_bnt_start_process(self):	
+		if (self.label_user.text() == ''):
+			msgBox = QMessageBox()
+			msgBox.setIcon(QMessageBox.Information)
+			msgBox.setText("Вы не вошли в аккаунт")
+			msgBox.setWindowTitle("Ошибка")
+			msgBox.setStandardButtons(QMessageBox.Ok)
+			msgBox.exec()
+		else:
+			result = ConnectToMySQL().RetriveBlobImg()
+			find = False
 
-		result = ConnectToMySQL().RetriveBlobImg()
-		find = False
+			self.appUI.label_11.setText("Идет распознавание")
 
-		self.appUI.label_11.setText("Идет распознавание")
+			try:
 
-		try:
+				for i in range(len(result)):
+					id = result[i]['User_id']
 
-			for i in range(len(result)):
-				id = result[i]['User_id']
+					with open(f"assets/tmp/recognition/tmp{id}.jpg", "wb") as fh:
+						fh.write(base64.decodebytes(result[i]['photo']))
 
-				with open(f"assets/tmp/recognition/tmp{id}.jpg", "wb") as fh:
-					fh.write(base64.decodebytes(result[i]['photo']))
+					if not os.path.exists('assets/tmp/recognition'):
+						print("[error] no serch directory")
+						sys.exit()
 
-				if not os.path.exists('assets/tmp/recognition'):
-					print("[error] no serch directory")
-					sys.exit()
-
-			know_encodings = []
-			images = os.listdir('assets/tmp/recognition')
-		
-			for (j, image) in enumerate(images):
-
-				face_img = face_recognition.load_image_file(f'assets/tmp/recognition/{image}')
-				face_enc = face_recognition.face_encodings(face_img)[0]
-
-				if len(know_encodings) == 0:
-					know_encodings.append(face_enc)
-				else:
-					for item in range(0, len(know_encodings)):
-						result = face_recognition.compare_faces([face_enc], know_encodings[item])
-
-						if result[0]:
-							know_encodings.append(face_enc)
-
-							id = re.sub("[t|m|p|.|j|p|g]", "", image)
-
-							db = ConnectToMySQL()
-							info = db.RetrieveOutBlob(id)
-
-							age = str(info['birthday']).split('-')[0]
-							age = int(age) - int(datetime.datetime.now().year)
-							age = re.sub("-", "", str(age))
-
-							self.appUI.label_11.setText('Присутствует в базе данны')
-							self.appUI.label_13.setText(info['fio'])
-							self.appUI.label_15.setText(info['pol'])
-							self.appUI.label_17.setText(str(info['birthday']))
-							self.appUI.label_19.setText(age)
-							self.appUI.label_21.setText(str(info['address']))
-
-							find = True
-							break
-						else:
-							self.appUI.label_11.setText('Такого человека нет в базе данных')
-							self.appUI.label_13.clear()
-							self.appUI.label_15.clear()
-							self.appUI.label_17.clear()
-							self.appUI.label_19.clear()
-							self.appUI.label_21.clear()
+				know_encodings = []
+				images = os.listdir('assets/tmp/recognition')
 			
-							break
-				if find:
-					break 
-		except Exception:
-			self.appUI.label_11.setText("Лицо было не найденно")
-			self.appUI.label_13.clear()
-			self.appUI.label_15.clear()
-			self.appUI.label_17.clear()
-			self.appUI.label_19.clear()
-			self.appUI.label_21.clear()
+				for (j, image) in enumerate(images):
 
-		finally:
-			for image in images:
-				print('удаленно фото', image)
-				os.remove(f'assets/tmp/recognition/{image}')
+					face_img = face_recognition.load_image_file(f'assets/tmp/recognition/{image}')
+					face_enc = face_recognition.face_encodings(face_img)[0]
+
+					if len(know_encodings) == 0:
+						know_encodings.append(face_enc)
+					else:
+						for item in range(0, len(know_encodings)):
+							result = face_recognition.compare_faces([face_enc], know_encodings[item])
+
+							if result[0]:
+								know_encodings.append(face_enc)
+
+								id = re.sub("[t|m|p|.|j|p|g]", "", image)
+
+								db = ConnectToMySQL()
+								info = db.RetrieveOutBlob(id)
+
+								age = str(info['birthday']).split('-')[0]
+								age = int(age) - int(datetime.datetime.now().year)
+								age = re.sub("-", "", str(age))
+
+								self.appUI.label_11.setText('Присутствует в базе данны')
+								self.appUI.label_13.setText(info['fio'])
+								self.appUI.label_15.setText(info['pol'])
+								self.appUI.label_17.setText(str(info['birthday']))
+								self.appUI.label_19.setText(age)
+								self.appUI.label_21.setText(str(info['address']))
+
+								find = True
+								break
+							else:
+								self.appUI.label_11.setText('Такого человека нет в базе данных')
+								self.appUI.label_13.clear()
+								self.appUI.label_15.clear()
+								self.appUI.label_17.clear()
+								self.appUI.label_19.clear()
+								self.appUI.label_21.clear()
+				
+								break
+					if find:
+						break 
+			except Exception:
+				self.appUI.label_11.setText("Лицо было не найденно")
+				self.appUI.label_13.clear()
+				self.appUI.label_15.clear()
+				self.appUI.label_17.clear()
+				self.appUI.label_19.clear()
+				self.appUI.label_21.clear()
+
+			finally:
+				for image in images:
+					print('удаленно фото', image)
+					os.remove(f'assets/tmp/recognition/{image}')
+
 	@pyqtSlot(bool)
 	def on_btn_del_connect(self):
 		date = [
@@ -477,26 +518,32 @@ class MainWindow(QtWidgets.QMainWindow):
 
 	@pyqtSlot(bool)
 	def on_btn_add_photo(self):
-		try:
-			label = self.appUI.label_upload_photo
-			photo = self.OpenFileDiolog()
-			self.appUI.status_label_2.setText(photo)
-			self.path = self.appUI.status_label_2.text()
-			shutil.copyfile(self.path, "assets/tmp/recognition/past_user_image.jpg")
-			label.setPixmap(QPixmap(photo).scaled(500, 690))
-			label.setScaledContents(True)
-		except Exception:
-			msgBox = QMessageBox()
-			msgBox.setIcon(QMessageBox.Information)
-			msgBox.setText("Вы не вставили фото")
-			msgBox.setWindowTitle("Ошибка")
-			msgBox.setStandardButtons(QMessageBox.Ok)
-			msgBox.exec()
-
-	@pyqtSlot(bool)
-	def on_btn_clear_table_1(self):
-		while (self.result_table_1.rowCount() > 0):
-			self.result_table_1.removeRow(0)
+		if (self.label_user.text() == ''):
+				msgBox = QMessageBox()
+				msgBox.setIcon(QMessageBox.Information)
+				msgBox.setText("Вы не вошли в аккаунт")
+				msgBox.setWindowTitle("Ошибка")
+				msgBox.setStandardButtons(QMessageBox.Ok)
+				msgBox.exec()
+		else:
+			try:
+				label = self.appUI.label_upload_photo
+				photo = self.OpenFileDiolog()
+				self.appUI.status_label_2.setText(photo)
+				self.path = self.appUI.status_label_2.text()
+				if not os.path.exists("assets/tmp/recognition"):
+					os.makedirs("assets/tmp/recognition") 
+				shutil.copyfile(self.path, "assets/tmp/recognition/past_user_image.jpg")
+				label.setPixmap((QPixmap(photo)))
+				# .scaled(490, 740)
+				# label.setScaledContents(True)
+			except Exception:
+				msgBox = QMessageBox()
+				msgBox.setIcon(QMessageBox.Information)
+				msgBox.setText("Вы не вставили фото")
+				msgBox.setWindowTitle("Ошибка")
+				msgBox.setStandardButtons(QMessageBox.Ok)
+				msgBox.exec()
 
 	@pyqtSlot(bool)
 	def on_btn_clear_table_2(self):
@@ -526,6 +573,8 @@ class MainWindow(QtWidgets.QMainWindow):
 		lineEdit_pol = self.appUI.lineEdit_pol.text()
 		lineEdit_data = self.appUI.lineEdit_data.text()
 		lineEdit_addres = self.appUI.lineEdit_addres.text()
+		lineEdit_admin = self.appUI.lineEdit_admin.text()
+		lineEdit_polsovatel = self.appUI.lineEdit_polsovatel.text()
 		label_status = self.appUI.label_status.text()
 		if (label_status == ''):
 			msgBox = QMessageBox()
@@ -534,7 +583,7 @@ class MainWindow(QtWidgets.QMainWindow):
 			msgBox.setWindowTitle("Ошибка")
 			msgBox.setStandardButtons(QMessageBox.Ok)
 			msgBox.exec()
-		elif (lineEdit_id == "" or lineEdit_password == "" or lineEdit_fio == "" or lineEdit_pol  == "" or lineEdit_data == "" or lineEdit_addres == ""):
+		elif (lineEdit_id == "" or lineEdit_password == "" or lineEdit_fio == "" or lineEdit_pol  == "" or lineEdit_data == "" or lineEdit_addres == "" or lineEdit_admin == "" or lineEdit_polsovatel == ""): 
 			msgBox = QMessageBox()
 			msgBox.setIcon(QMessageBox.Information)
 			msgBox.setText("Вы не заполнили все строки данными")
@@ -543,31 +592,39 @@ class MainWindow(QtWidgets.QMainWindow):
 			msgBox.exec()
 		else:
 			db = ConnectToMySQL()
-			db.UploadData(lineEdit_id, lineEdit_login, lineEdit_password, lineEdit_fio, lineEdit_pol, lineEdit_data, lineEdit_addres, label_status)
+			db.UploadData(lineEdit_id, lineEdit_login, lineEdit_password, lineEdit_fio, lineEdit_pol, lineEdit_data, lineEdit_addres, label_status, lineEdit_admin, lineEdit_polsovatel)
 
 	@pyqtSlot(bool)
 	def on_btn_get_date_1(self):
-		result = ConnectToMySQL().get_all_data_from_db()
+		if (self.label_user.text() == ''):
+			msgBox = QMessageBox()
+			msgBox.setIcon(QMessageBox.Information)
+			msgBox.setText("Вы не вошли в аккаунт")
+			msgBox.setWindowTitle("Ошибка")
+			msgBox.setStandardButtons(QMessageBox.Ok)
+			msgBox.exec()
+		else:
+			result = ConnectToMySQL().get_all_data_from_db()
 
-		if result:
-			self.result_table_1.setRowCount(len(result))
+			if result:
+				self.result_table_1.setRowCount(len(result))
 
-			for row, item in enumerate(result):
-				column_1_item = QtWidgets.QTableWidgetItem(str(item['id']))
-				column_2_item = QtWidgets.QTableWidgetItem(str(item['login']))
-				column_3_item = QtWidgets.QTableWidgetItem(str(item['fio']))
-				column_4_item = QtWidgets.QTableWidgetItem(str(item['pol']))
-				column_5_item = QtWidgets.QTableWidgetItem(str(item['birthday']))
-				column_6_item = QtWidgets.QTableWidgetItem(str(item['address']))
-				column_7_item = QtWidgets.QTableWidgetItem(str('view'))			
+				for row, item in enumerate(result):
+					column_1_item = QtWidgets.QTableWidgetItem(str(item['id']))
+					column_2_item = QtWidgets.QTableWidgetItem(str(item['login']))
+					column_3_item = QtWidgets.QTableWidgetItem(str(item['fio']))
+					column_4_item = QtWidgets.QTableWidgetItem(str(item['pol']))
+					column_5_item = QtWidgets.QTableWidgetItem(str(item['birthday']))
+					column_6_item = QtWidgets.QTableWidgetItem(str(item['address']))
+					column_7_item = QtWidgets.QTableWidgetItem(str('view'))			
 
-				self.result_table_1.setItem(row, 0, column_1_item)
-				self.result_table_1.setItem(row, 1, column_2_item)
-				self.result_table_1.setItem(row, 2, column_3_item)
-				self.result_table_1.setItem(row, 3, column_4_item)
-				self.result_table_1.setItem(row, 4, column_5_item)
-				self.result_table_1.setItem(row, 5, column_6_item)
-				self.result_table_1.setItem(row, 6, column_7_item)
+					self.result_table_1.setItem(row, 0, column_1_item)
+					self.result_table_1.setItem(row, 1, column_2_item)
+					self.result_table_1.setItem(row, 2, column_3_item)
+					self.result_table_1.setItem(row, 3, column_4_item)
+					self.result_table_1.setItem(row, 4, column_5_item)
+					self.result_table_1.setItem(row, 5, column_6_item)
+					self.result_table_1.setItem(row, 6, column_7_item)
 
 	@pyqtSlot(bool)
 	def on_btn_get_date_2(self):
